Prime numbers
A natural number is prime if it is greater than 1 and has no divisions other than 1 and itself. I wrote a function checking whether a number is prime or not using Python. I used a range from 1 to 20, I used one argument to check the value if it returns as True, then my argument is a prime number, and False otherwise.